master edit start
