+++ app/src/main/java/com/example/sharemap2/PersonFragment.java
@@ -1,0 +1,25 @@
+package com.example.sharemap2;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+
+
+public class PersonFragment extends Fragment {
+
+
+    //Hellow World
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_person, container, false);
+    }
+
+
+}
