--- app/src/main/java/com/example/sharemap2/fragment_search_root.java
@@ -1,38 +1,0 @@
-package com.example.sharemap2;
-
-import android.content.Context;
-import android.location.LocationListener;
-import android.location.LocationManager;
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import com.google.android.gms.location.FusedLocationProviderClient;
-import com.google.android.gms.location.LocationServices;
-
-import org.w3c.dom.Text;
-
-
-public class fragment_search_root extends Fragment {
-
-
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View view=inflater.inflate(R.layout.fragment_search, container, false);
-
-        // Inflate the layout for this fragment
-
-        return view;
-    }
-
-
-
-}
