--- app/src/main/res/layout/fragment_upload.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragment_upload_root">
-
-
-
-
-    <fragment
-            android:id="@+id/map"
-            android:name="com.google.android.gms.maps.SupportMapFragment"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            tools:context=".fragment_upload_root"/>
-
-    <Button
-        android:id="@+id/button"
-        android:layout_width="58dp"
-        android:layout_height="57dp"
-        android:layout_marginStart="50dp"
-        android:layout_marginTop="500dp"
-        android:background="@drawable/ic_add_circle_black_24dp"
-        app:layout_constraintVertical_bias="0.862" />
-
-    <Button
-        android:id="@+id/button2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="250dp"
-        android:layout_marginTop="500dp"
-        android:text="upload" />
-
-    <TextView
-        android:id="@+id/textView1"
-        android:layout_width="match_parent"
-        android:layout_height="30dp"
-        android:background="#FFFFFF"
-        android:text="TextView"
-        android:textSize="20sp" />
-
-
-</FrameLayout>
