+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="title_activity_root_edit">RootEditActivity</string>
 </resources>
